Use Retry for for retry_on_transient_error
Use the same core retry object for the retry wrapper. Update defaults to
match the previous expected result.
