+++ lib/spack/spack/test/oci/urlopen.py
@@ -743,14 +743,11 @@ class BrokenServer(DummyServer):
         ("https://example.com/not-found/", 3, True, 1),
     ],
 )
-def test_retry(url, max_retries, expect_failure, expect_requests):
+def test_retry(url, max_retries, expect_failure, expect_requests, mock_sleep):
     server = BrokenServer("example.com")
     urlopen = create_opener(server).open
-    sleep_time = []
-    dont_sleep = lambda t: sleep_time.append(t)  # keep track of sleep times
-
     try:
-        response = default_retry(urlopen, retries=max_retries, sleep=dont_sleep)(url)
+        response = default_retry(urlopen, spack.util.web.Retry(total=max_retries))(url)
     except urllib.error.HTTPError as e:
         if not expect_failure:
             assert False, f"Unexpected HTTPError: {e}"
@@ -760,7 +757,7 @@ def test_retry(url, max_retries, expect_failure, expect_requests):
         assert response.status == 200
 
     assert len(server.requests) == expect_requests
-    assert sleep_time == [2**i for i in range(expect_requests - 1)]
+    assert mock_sleep.times == [2**i for i in range(expect_requests - 1)]
 
 
 def test_list_tags():
